Integrate ts files into sfc
