+++ pages/actors/index.vue
@@ -17,7 +17,28 @@
   </div>
 </template>
 
-<script lang="ts" src="./index.ts">
+<script lang="ts">
+import ActorIcon from '~/components/ActorIcon.vue'
+
+export default {
+  components: {
+    ActorIcon
+  },
+  computed: {
+    sortedActors() {
+      return [...this.$store.getters.actorsWithEpisodes].sort((a, b) => {
+        // sort by appearCount or lastAppearDate
+        const countDiff = b.episodes.length - a.episodes.length
+        return countDiff !== 0 ? countDiff : b.episodes[0].published - a.episodes[0].published
+      })
+    }
+  },
+  head() {
+    return {
+      title: 'All actors'
+    }
+  }
+}
 </script>
 <style lang="scss" scoped>
 .actor-list {
