+++ pages/episode.vue
@@ -4,10 +4,10 @@
 
 <script>
 export default {
-  transition (to, from) {
-    return (!from || from.name !== to.name)
-    ? 'page'
-    : Number(to.params.id) < Number(from.params.id) ? 'slide-right' : 'slide-left'
+  transition(to, from) {
+    return !from || from.name !== to.name
+      ? 'page'
+      : Number(to.params.id) < Number(from.params.id) ? 'slide-right' : 'slide-left'
   }
 }
 </script>
